# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type LanguageDirection = typeof LanguageDirection[keyof typeof LanguageDirection];


export const LanguageDirection = {
  ltr: 'ltr',
  rtl: 'rtl',
} as const;

export interface Language {
  code: string;
  label: string;
  nativeLabel: string;
  direction: LanguageDirection;
}

export interface AppConfig {
  serviceConfigured: boolean;
  model: string;
  maxImageBytes: number;
  maxChapterBytes: number;
  supportedLanguages: Language[];
}

export interface ImageTranslationInput {
  file: Blob;
  targetLanguage: string;
}

export interface ChapterTranslationInput {
  file?: Blob;
  images?: Blob[];
  targetLanguage: string;
}

export interface MultiChapterTranslationInput {
  /** @maxItems 3 */
  files: Blob[];
  targetLanguage: string;
}

export type TextRegionAlign = typeof TextRegionAlign[keyof typeof TextRegionAlign];


export const TextRegionAlign = {
  left: 'left',
  center: 'center',
  right: 'right',
} as const;

export interface TextRegion {
  text: string;
  translatedText: string;
  x: number;
  y: number;
  width: number;
  height: number;
  orientation: string;
  regionType: string;
  confidence: number;
  fontSize: number;
  align: TextRegionAlign;
}

export interface TranslationResult {
  id: string;
  originalImage: string;
  translatedImage: string;
  width: number;
  height: number;
  regions: TextRegion[];
  targetLanguage: string;
  model: string;
}

export interface ChapterPage {
  pageNumber: number;
  originalImage: string;
  translatedImage: string;
  width: number;
  height: number;
  regions: TextRegion[];
}

export interface ChapterResult {
  id: string;
  name: string;
  pageCount: number;
  pages: ChapterPage[];
  targetLanguage: string;
  model: string;
}

export interface MultiChapterResult {
  id: string;
  chapterCount: number;
  chapters: ChapterResult[];
  targetLanguage: string;
}