# lib/api-zod/src/generated/types/appConfig.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Language } from './language';

export interface AppConfig {
  serviceConfigured: boolean;
  model: string;
  maxImageBytes: number;
  maxChapterBytes: number;
  supportedLanguages: Language[];
}