# lib/api-zod/src/generated/types/chapterPage.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TextRegion } from './textRegion';

export interface ChapterPage {
  pageNumber: number;
  originalImage: string;
  translatedImage: string;
  width: number;
  height: number;
  regions: TextRegion[];
}