# lib/api-zod/src/generated/types/chapterResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChapterPage } from './chapterPage';

export interface ChapterResult {
  id: string;
  name: string;
  pageCount: number;
  pages: ChapterPage[];
  targetLanguage: string;
  model: string;
}