# lib/api-zod/src/generated/types/language.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LanguageDirection } from './languageDirection';

export interface Language {
  code: string;
  label: string;
  nativeLabel: string;
  direction: LanguageDirection;
}