# lib/api-zod/src/generated/types/multiChapterResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChapterResult } from './chapterResult';

export interface MultiChapterResult {
  id: string;
  chapterCount: number;
  chapters: ChapterResult[];
  targetLanguage: string;
}