# lib/api-zod/src/generated/types/multiChapterTranslationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MultiChapterTranslationInput {
  /** @maxItems 3 */
  files: Blob[];
  targetLanguage: string;
}