# lib/api-zod/src/generated/types/textRegion.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TextRegionAlign } from './textRegionAlign';

export interface TextRegion {
  text: string;
  translatedText: string;
  x: number;
  y: number;
  width: number;
  height: number;
  orientation: string;
  regionType: string;
  confidence: number;
  fontSize: number;
  align: TextRegionAlign;
}